ENH: Truncate negative dividend adjustment ratios to 0.05
Occasionally assets pay dividends greater than the last traded
value of the asset. This can occur when investors are uncertain
about the nature of the dividend payments, which means they are
not able to correctly price the asset in advance of the dividend.
After these payments the assets fall dramatically in value. Use
a 5% adjustment factor as it is consistent with the case of VXUP
and VXDN: http://finance.yahoo.com/q?s=VXUP.